+++ plugin-dev/Source/Sentry/Private/GenericPlatform/GenericPlatformSentrySubsystem.h
@@ -177,11 +177,15 @@ private:
 	FThreadSafeBool bIsCrashing;
 
 #ifdef USE_SENTRY_SESSION_REPLAY
+	void StartSessionReplay(const USentrySettings* settings);
+
 	FString GetReplayPath() const;
 
 	FString SessionReplayId;
 
 	TUniquePtr<FSentrySessionReplayRecorder> SessionReplay;
+
+	FDelegateHandle EngineLoopInitCompleteHandle;
 #endif
 };